Resolving MSVC 2017 compiler warnings

+++ src/BabylonCpp/include/babylon/animations/animatable.h
@@ -145,8 +145,8 @@ protected:
    * is not looping
    * @param animations defines a group of animation to add to the new Animatable
    */
-  Animatable(Scene* scene, const IAnimatablePtr& target, int fromFrame = 0,
-             int toFrame = 100, bool loopAnimation = false,
+  Animatable(Scene* scene, const IAnimatablePtr& target, float fromFrame = 0.f,
+             float toFrame = 100.f, bool loopAnimation = false,
              float speedRatio                            = 1.f,
              const std::function<void()>& onAnimationEnd = nullptr,
              const std::vector<AnimationPtr>& animations = {});
@@ -207,12 +207,12 @@ public:
   /**
    * Defines the starting frame number (default is 0)
    */
-  int fromFrame;
+  float fromFrame;
 
   /**
    * Defines the ending frame number (default is 100)
    */
-  int toFrame;
+  float toFrame;
 
   /**
    * Defines if the animation must loop (default is false)
